GUI improved, now supports one-time server data fetching -> reading serverLists and joining chat-rooms
GUI needs to be cleaned up, remove unnecessary variables and create
methods. separate network manager code from GUI code.

Database needs to be implemented for
	-messsage back up of around 30 messages so users can see brief chat
history when they join
	-username and passwords for log in authentication

polling for data fetching needs to be implemented to allow for
serverLists, participantsLists etc to be constantly up to date.

some method to create and delete chatrooms

overall clean up and restructuring for organization.